working separated profilewrapper
